Ask for all fields from the content provider.
It is not a particularly expensive operation in the scheme of things. When we
are going to the database, the bottlneck is in disk access for the actual query
of the database tables. The difference between retrieving two columns or
all the columns when the query is for a handful of apps is inconsequential.

Thus, it is better to be safe than sorry and just ask for all the things so
that our value objects are correctly populated.