only show update notification if updates are going to happen
UpdateService.onHandleIntent() starts with a time check for whether an
update is actually scheduled.  Before, UpdateService put up a notification
when it started.  This changes it so that the notification is put up after
the check, so it should only show the notification if UpdateService is
actually going to run, and no longer when it is just waking up to check the
time.

!307 #662